boot: move reading efi vars to bootloader/efi pkg, export
This splits up the boot package function to get the partition uuid for the
booted kernel into the main boot package function logic, and specifics of
reading EFI variables in the bootloader package.
